refactor: changement du chemin du dossier store
