+++ apps/front/src/modules/job-postings/job-postings.svelte
@@ -4,7 +4,7 @@
 
 <script>
     import { Link } from 'svelte-routing';
-    import { JobPostingsStore } from '../../stores';
+    import { JobPostingsStore } from '../../store';
     import { onMount } from 'svelte';
 
     onMount(() => JobPostingsStore.init());
