+++ apps/front/src/store/job-posting.js
