+++ apps/front/src/store/job-postings.js
